fix : change Long validating Logic position

+++ BE/src/main/java/roomescape/presentation/ReservationController.java
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import roomescape.application.ReservationService;
 import roomescape.entity.Reservation;
+import roomescape.global.exception.ErrorCode;
+import roomescape.global.exception.customException.ReservationException;
 import roomescape.presentation.dto.AvailableReservationResponse;
 import roomescape.presentation.dto.ReservationRequest;
 import roomescape.presentation.dto.ReservationResponse;
@@ -82,6 +84,9 @@ public class ReservationController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
+        if (id == null) {
+            throw new ReservationException(ErrorCode.RESERVATION_ID_NULL);
+        }
         service.deleteById(id);
         return ResponseEntity.noContent().build();
     }
